Fixed number of steps

+++ AlgoritimosDeOrdenacao/src/Global.cpp
@@ -2,7 +2,7 @@
 
 bool isRunning = false;
 float lastExecutionTime = 0.0;
-int lastNumberOfSteps = 0;
+unsigned long int lastNumberOfSteps = 0;
 double deltaDecaSps = 0;
 
 void Global::alert(std::string text)
